+++ CHANGELOG.md
@@ -15,6 +15,8 @@ across Flow
 
 * Now you can receive notifications for pending transactions on iOS and Android.
 It may support macOS in the future.
+* Deleted transaction now go to "Trash bin". You can recover recently deleted
+items, closes [#294](https://github.com/flow-mn/flow/issues/294)
 
 ## Beta 0.11.1